Refactor of code and move to single line the run line.
