+++ package.json
@@ -4,9 +4,10 @@
   "description": "Server side rendering project using React",
   "main": "index.js",
   "scripts": {
+    "dev": "npm-run-all --parallel dev:*",
     "dev:server": "nodemon --watch build --exec \"node build/bundle.js\"",
-    "dev:build:server": "webpack --config webpack.server.js --watch",
-    "dev:build:client": "webpack --config webpack.client.js --watch"
+    "dev:build-server": "webpack --config webpack.server.js --watch",
+    "dev:build-client": "webpack --config webpack.client.js --watch"
   },
   "author": "@miguelmejiaq",
   "license": "ISC",
