+++ src/index.js
@@ -1,24 +1,10 @@
 import express from 'express';
-import React from 'react';
-import {renderToString} from 'react-dom/server';
-import Home from './client/components/Home';
+import renderer from './helpers/renderer';
 
 const app = express();
 app.use(express.static('public'));
 app.get('/', (req,res) => {
-    const content = renderToString(<Home />);
-    const html= `
-        <html>
-            <head></head>
-            <body>
-                <div id="root">
-                ${content}
-                </div>
-                <script src="bundle.js"><script>
-            </body>
-        </html/>
-    `;
-    res.send(html);
+    res.send(renderer());
 });
 
 app.listen(3000, () => {
