+++ webpack.base.js
@@ -1,0 +1,19 @@
+module.exports = {
+    // Tell webpack to run babel on every js file
+    module: {
+        rules: [
+            {
+                test: /\.js?$/,
+                loader: 'babel-loader',
+                exclude: /node_modules/,
+                options: {
+                    presets: [
+                        'react',
+                        'stage-0',
+                        ['env', { targets: { browsers: ['last 2 versions'] }}]
+                    ]
+                }
+            }
+        ]
+    }
+}
