+++ webpack.client.js
@@ -1,5 +1,7 @@
 const path = require('path');
-module.exports = {
+const merge = require('webpack-merge');
+const baseConfig = require('./webpack.base.js')
+const config = {
     // Tell webpack the root file of out
     // server application
     entry: './src/client/client.js',
@@ -7,22 +9,6 @@ module.exports = {
     output: {
         filename: 'bundle.js',
         path: path.resolve(__dirname,'public')
-    },
-    // Tell webpack to run babel on every js file
-    module: {
-        rules: [
-            {
-                test: /\.js?$/,
-                loader: 'babel-loader',
-                exclude: /node_modules/,
-                options: {
-                    presets: [
-                        'react',
-                        'stage-0',
-                        ['env', { targets: { browsers: ['last 2 versions'] }}]
-                    ]
-                }
-            }
-        ]
     }
-}
+};
+module.exports = merge(baseConfig, config);
