+++ webpack.server.js
@@ -1,5 +1,8 @@
 const path = require('path');
-module.exports = {
+const merge = require('webpack-merge');
+const baseConfig = require('./webpack.base.js')
+const webpackNodeExternals = require('webpack-node-externals');
+const config = {
     // Inform webpack that we're building a bundle 
     // for Node JS, rather than for the browser
     target: 'node',
@@ -11,21 +14,6 @@ module.exports = {
         filename: 'bundle.js',
         path: path.resolve(__dirname,'build')
     },
-    // Tell webpack to run babel on every js file
-    module: {
-        rules: [
-            {
-                test: /\.js?$/,
-                loader: 'babel-loader',
-                exclude: /node_modules/,
-                options: {
-                    presets: [
-                        'react',
-                        'stage-0',
-                        ['env', { targets: { browsers: ['last 2 versions'] }}]
-                    ]
-                }
-            }
-        ]
-    }
-}
+    externals: [webpackNodeExternals()]
+};
+module.exports = merge(baseConfig, config);
